[!!] fixed things in 3D [++] added clincal NN model with training alg [++] added combination model + prepared for combination experiments

+++ Training/Utils/cnn_utils.py
@@ -168,11 +168,12 @@ def _conv_layer_pure_3d(input, shape, strides=[1, 1, 1, 1, 1], padding='SAME'):
 		'weights',
 		shape,
 		initializer=tf.contrib.layers.variance_scaling_initializer(),
-		dtype=tf.float16)
+		dtype=tf.float32)
 
 	conv = tf.nn.conv3d(input, kernel, strides=strides, padding=padding, name='Pre-Activation')
 
 	return conv, kernel
+
 
 def _conv_layer_3d_with_kernel(input, kernel, shape, strides, padding, is_training, bnorm=False):
 	# No bias when BN
